Moved generic components to more generic location.
