+++ src/views/common/Animation/AnimationControlButtonStyles.css
