+++ src/views/common/Animation/AnimationPlaybackControls.tsx
