+++ src/views/common/Animation/AnimationStateControlButtons.tsx
