+++ src/views/common/Animation/AnimationStatePlaybackBarLayer.tsx
