+++ src/views/common/Animation/AnimationStateReducer.tsx
